Updated the send message endpoint to upload files
